Trait lightning::api::distribution::DataDistribution

pub trait DataDistribution: Debug + AsAny + Send + Sync + 'static {
    fn visit_unique(
        &self, 
        region: Rect, 
        affinity: Option<MemoryId>, 
        callback: &mut dyn FnMut(ChunkQueryResult)
    );

    fn visit_replicated(
        &self, 
        region: Rect, 
        callback: &mut dyn FnMut(ChunkQueryResult)
    );

    fn as_work_distribution(&self) -> Option<&dyn WorkDistribution> { ... }

    fn clone_region(
        &self, 
        _system: &SystemInfo, 
        _region: Rect
    ) -> Result<(Arc<dyn DataDistribution>, Vec<ChunkDescriptor>)> { ... }
}

Interface that specifies how the data of an array should be distributed across the workers.

Required methods

fn visit_unique(
&self,
region: Rect,
affinity: Option<MemoryId>,
callback: &mut dyn FnMut(ChunkQueryResult)
)

Query how the data within the given region is distributed.

The provided callback will be called for each returned subregion. The union of the subregions represents the entire region. Subregions are disjoint (do not overlap), meaning each point in the given region will represented by one unique subregion.

The affinity parameter is used to signal that it is preferred if the chunks have affinity to the given memory space.

fn visit_replicated(
&self,
region: Rect,
callback: &mut dyn FnMut(ChunkQueryResult)
)

Query how the data within the given region is distributed.

The provided callback will be called for each returned subregion within the given region. The union of the subregions represents the entire region. All chunks that lie within the given region will be returned, meaning that the same point may be presented in multiple subregions if it is replicated.

Provided methods

fn as_work_distribution(&self) -> Option<&dyn WorkDistribution>

fn clone_region(
&self,
_system: &SystemInfo,
_region: Rect
) -> Result<(Arc<dyn DataDistribution>, Vec<ChunkDescriptor>)>