Trait lightning::api::distribution::IntoDataDistribution

pub trait IntoDataDistribution {
    fn into_data_distribution(
        self, 
        system: &SystemInfo, 
        size: Dim
    ) -> Result<(Arc<dyn DataDistribution>, Vec<ChunkDescriptor>)>;
}

Types that can be converted into an DataDistribution.

Anything implementing IntoDataDistribution can be converted into a concrete DataDistribution given an array size and the list of available workers. The trait could thus be seen as a “factory” object which can produce DataDistributions.

The difference between DataDistribution and IntoDataDistribution is that the first specifies a concrete data distribution for physical array, while the latter specifies an abstract description of a data distribution (for example: row-wise, column-wise, tiled, replicated, etc.)

Required methods

fn into_data_distribution(
self,
system: &SystemInfo,
size: Dim
) -> Result<(Arc<dyn DataDistribution>, Vec<ChunkDescriptor>)>