Module lightning::chain

Structs and traits which allow other parts of rust-lightning to interact with the blockchain.

Modules

chaininterface

Traits and utility impls which allow other parts of rust-lightning to interact with the blockchain.

chainmonitor

Logic to connect off-chain channel management with on-chain transaction monitoring.

channelmonitor

The logic to monitor for on-chain transactions and create the relevant claim responses lives here.

keysinterface

Provides keys to LDK and defines some useful objects describing spendable on-chain outputs.

transaction

Types describing on-chain transactions.

Structs

BestBlock

The best known block as identified by its hash and height.

WatchedOutput

A transaction output watched by a ChannelMonitor for spends on-chain.

Enums

AccessError

An error when accessing the chain via Access.

ChannelMonitorUpdateStatus

An enum representing the status of a channel monitor update persistence.

Traits

Access

The Access trait defines behavior for accessing chain data and state, such as blocks and UTXOs.

Confirm

The Confirm trait is used to notify LDK when relevant transactions have been confirmed on chain or unconfirmed during a chain reorganization.

Filter

The Filter trait defines behavior for indicating chain activity of interest pertaining to channels.

Listen

The Listen trait is used to notify when blocks have been connected or disconnected from the chain.

Watch

The Watch trait defines behavior for watching on-chain activity pertaining to channels as blocks are connected and disconnected.