Module lightning::chain::chaininterface

Traits and utility impls which allow other parts of rust-lightning to interact with the blockchain.

Includes traits for monitoring and receiving notifications of new blocks and block disconnections, transaction broadcasting, and feerate information requests.

Enums

ConfirmationTarget

An enum that represents the speed at which we want a transaction to confirm used for feerate estimation.

Constants

FEERATE_FLOOR_SATS_PER_KW

Minimum feerate that takes a sane approach to bitcoind weight-to-vbytes rounding. See the following Core Lightning commit for an explanation: https://github.com/ElementsProject/lightning/commit/2e687b9b352c9092b5e8bd4a688916ac50b44af0

MIN_RELAY_FEE_SAT_PER_1000_WEIGHT

Minimum relay fee as required by bitcoin network mempool policy.

Traits

BroadcasterInterface

An interface to send a transaction to the Bitcoin network.

FeeEstimator

A trait which should be implemented to provide feerate information on a number of time horizons.