Struct lightning::chain::chainmonitor::ChainMonitor

pub struct ChainMonitor<ChannelSigner: WriteableEcdsaChannelSigner, C: Deref, T: Deref, F: Deref, L: Deref, P: Deref>where
    C::Target: Filter,
    T::Target: BroadcasterInterface,
    F::Target: FeeEstimator,
    L::Target: Logger,
    P::Target: Persist<ChannelSigner>,{ /* private fields */ }

An implementation of chain::Watch for monitoring channels.

Connected and disconnected blocks must be provided to ChainMonitor as documented by chain::Watch. May be used in conjunction with ChannelManager to monitor channels locally or used independently to monitor channels remotely. See the module-level documentation for details.

Implementations

impl<ChannelSigner: WriteableEcdsaChannelSigner, C: Deref, T: Deref, F: Deref, L: Deref, P: Deref> ChainMonitor<ChannelSigner, C, T, F, L, P>where
C::Target: Filter,
T::Target: BroadcasterInterface,
F::Target: FeeEstimator,
L::Target: Logger,
P::Target: Persist<ChannelSigner>,

pub fn new(
chain_source: Option<C>,
broadcaster: T,
logger: L,
feeest: F,
persister: P
) -> Self

Creates a new ChainMonitor used to watch on-chain activity pertaining to channels.

When an optional chain source implementing chain::Filter is provided, the chain monitor will call back to it indicating transactions and outputs of interest. This allows clients to pre-filter blocks or only fetch blocks matching a compact filter. Otherwise, clients may always need to fetch full blocks absent another means for determining which blocks contain transactions relevant to the watched channels.

pub fn get_claimable_balances(
&self,
ignored_channels: &[&ChannelDetails]
) -> Vec<Balance>

Gets the balances in the contained ChannelMonitors which are claimable on-chain or claims which are awaiting confirmation.

Includes the balances from each ChannelMonitor except those included in ignored_channels, allowing you to filter out balances from channels which are still open (and whose balance should likely be pulled from the ChannelDetails).

See ChannelMonitor::get_claimable_balances for more details on the exact criteria for inclusion in the return value.

pub fn get_monitor(
&self,
funding_txo: OutPoint
) -> Result<LockedChannelMonitor<'_, ChannelSigner>, ()>

Gets the LockedChannelMonitor for a given funding outpoint, returning an Err if no such ChannelMonitor is currently being monitored for.

Note that the result holds a mutex over our monitor set, and should not be held indefinitely.

pub fn list_monitors(&self) -> Vec<OutPoint>

Lists the funding outpoint of each ChannelMonitor being monitored.

Note that ChannelMonitors are not removed when a channel is closed as they are always monitoring for on-chain state resolutions.

pub fn list_pending_monitor_updates(
&self
) -> HashMap<OutPoint, Vec<MonitorUpdateId>>

Lists the pending updates for each ChannelMonitor (by OutPoint being monitored).

pub fn channel_monitor_updated(
&self,
funding_txo: OutPoint,
completed_update_id: MonitorUpdateId
) -> Result<(), APIError>

Indicates the persistence of a ChannelMonitor has completed after ChannelMonitorUpdateStatus::InProgress was returned from an update operation.

Thus, the anticipated use is, at a high level:

This ChainMonitor calls Persist::update_persisted_channel which stores the update to disk and begins updating any remote (e.g. watchtower/backup) copies, returning ChannelMonitorUpdateStatus::InProgress,

once all remote copies are updated, you call this function with the completed_update_id that completed, and once all pending updates have completed the channel will be re-enabled.

Returns an APIError::APIMisuseError if funding_txo does not match any currently registered ChannelMonitors.

pub async fn process_pending_events_async<Future: Future, H: Fn(Event) -> Future>(
&self,
handler: H
)

Processes any events asynchronously in the order they were generated since the last call using the given event handler.

See the trait-level documentation of EventsProvider for requirements.