Struct lightning::chain::keysinterface::DelayedPaymentOutputDescriptor

pub struct DelayedPaymentOutputDescriptor {
    pub outpoint: OutPoint,
    pub per_commitment_point: PublicKey,
    pub to_self_delay: u16,
    pub output: TxOut,
    pub revocation_pubkey: PublicKey,
    pub channel_keys_id: [u8; 32],
    pub channel_value_satoshis: u64,
}

Information about a spendable output to a P2WSH script.

See SpendableOutputDescriptor::DelayedPaymentOutput for more details on how to spend this.

Fields
outpoint: OutPoint

The outpoint which is spendable.

per_commitment_point: PublicKey

Per commitment point to derive the delayed payment key by key holder.

to_self_delay: u16

The nSequence value which must be set in the spending input to satisfy the OP_CSV in the witness_script.

output: TxOut

The output which is referenced by the given outpoint.

revocation_pubkey: PublicKey

The revocation point specific to the commitment transaction which was broadcast. Used to derive the witnessScript for this output.

channel_keys_id: [u8; 32]

Arbitrary identification information returned by a call to ChannelSigner::channel_keys_id. This may be useful in re-deriving keys used in the channel to spend the output.

channel_value_satoshis: u64

The value of the channel which this output originated from, possibly indirectly.

Implementations

impl DelayedPaymentOutputDescriptor

pub const MAX_WITNESS_LENGTH: usize = 153usize

The maximum length a well-formed witness spending one of these should have.