Struct lightning::chain::keysinterface::InMemorySigner

pub struct InMemorySigner {
    pub funding_key: SecretKey,
    pub revocation_base_key: SecretKey,
    pub payment_key: SecretKey,
    pub delayed_payment_base_key: SecretKey,
    pub htlc_base_key: SecretKey,
    pub commitment_seed: [u8; 32],
    /* private fields */
}

A simple implementation of WriteableEcdsaChannelSigner that just keeps the private keys in memory.

This implementation performs no policy checks and is insufficient by itself as a secure external signer.

Fields
funding_key: SecretKey

Holder secret key in the 2-of-2 multisig script of a channel. This key also backs the holder’s anchor output in a commitment transaction, if one is present.

revocation_base_key: SecretKey

Holder secret key for blinded revocation pubkey.

payment_key: SecretKey

Holder secret key used for our balance in counterparty-broadcasted commitment transactions.

delayed_payment_base_key: SecretKey

Holder secret key used in an HTLC transaction.

htlc_base_key: SecretKey

Holder HTLC secret key used in commitment transaction HTLC outputs.

commitment_seed: [u8; 32]

Commitment seed.

Implementations

impl InMemorySigner

pub fn new<C: Signing>(
secp_ctx: &Secp256k1<C>,
funding_key: SecretKey,
revocation_base_key: SecretKey,
payment_key: SecretKey,
delayed_payment_base_key: SecretKey,
htlc_base_key: SecretKey,
commitment_seed: [u8; 32],
channel_value_satoshis: u64,
channel_keys_id: [u8; 32]
) -> InMemorySigner

Creates a new InMemorySigner.

pub fn counterparty_pubkeys(&self) -> &ChannelPublicKeys

Returns the counterparty’s pubkeys.

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn counterparty_selected_contest_delay(&self) -> u16

Returns the contest_delay value specified by our counterparty and applied on holder-broadcastable transactions, i.e., the amount of time that we have to wait to recover our funds if we broadcast a transaction.

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn holder_selected_contest_delay(&self) -> u16

Returns the contest_delay value specified by us and applied on transactions broadcastable by our counterparty, i.e., the amount of time that they have to wait to recover their funds if they broadcast a transaction.

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn is_outbound(&self) -> bool

Returns whether the holder is the initiator.

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn funding_outpoint(&self) -> &OutPoint

Funding outpoint

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn get_channel_parameters(&self) -> &ChannelTransactionParameters

Returns a ChannelTransactionParameters for this channel, to be used when verifying or building transactions.

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn opt_anchors(&self) -> bool

Returns whether anchors should be used.

Will panic if ChannelSigner::provide_channel_parameters has not been called before.

pub fn sign_counterparty_payment_input<C: Signing>(
&self,
spend_tx: &Transaction,
input_idx: usize,
descriptor: &StaticPaymentOutputDescriptor,
secp_ctx: &Secp256k1<C>
) -> Result<Vec<Vec<u8>>, ()>

Sign the single input of spend_tx at index input_idx, which spends the output described by descriptor, returning the witness stack for the input.

Returns an error if the input at input_idx does not exist, has a non-empty script_sig, is not spending the outpoint described by descriptor.outpoint, or if an output descriptor script_pubkey does not match the one we can spend.

pub fn sign_dynamic_p2wsh_input<C: Signing>(
&self,
spend_tx: &Transaction,
input_idx: usize,
descriptor: &DelayedPaymentOutputDescriptor,
secp_ctx: &Secp256k1<C>
) -> Result<Vec<Vec<u8>>, ()>

Sign the single input of spend_tx at index input_idx which spends the output described by descriptor, returning the witness stack for the input.

Returns an error if the input at input_idx does not exist, has a non-empty script_sig, is not spending the outpoint described by descriptor.outpoint, does not have a sequence set to descriptor.to_self_delay, or if an output descriptor script_pubkey does not match the one we can spend.