Struct lightning::chain::keysinterface::PhantomKeysManager

pub struct PhantomKeysManager { /* private fields */ }

Similar to KeysManager, but allows the node using this struct to receive phantom node payments.

A phantom node payment is a payment made to a phantom invoice, which is an invoice that can be paid to one of multiple nodes. This works because we encode the invoice route hints such that LDK will recognize an incoming payment as destined for a phantom node, and collect the payment itself without ever needing to forward to this fake node.

Phantom node payments are useful for load balancing between multiple LDK nodes. They also provide some fault tolerance, because payers will automatically retry paying other provided nodes in the case that one node goes down.

Note that multi-path payments are not supported in phantom invoices for security reasons. Switching between this struct and KeysManager will invalidate any previously issued invoices and attempts to pay previous invoices will fail.

Implementations

impl PhantomKeysManager

pub fn new(
seed: &[u8; 32],
starting_time_secs: u64,
starting_time_nanos: u32,
cross_node_seed: &[u8; 32]
) -> Self

Constructs a PhantomKeysManager given a 32-byte seed and an additional cross_node_seed that is shared across all nodes that intend to participate in phantom node payments together.

See KeysManager::new for more information on seed, starting_time_secs, and starting_time_nanos.

cross_node_seed must be the same across all phantom payment-receiving nodes and also the same across restarts, or else inbound payments may fail.

pub fn spend_spendable_outputs<C: Signing>(
&self,
descriptors: &[&SpendableOutputDescriptor],
outputs: Vec<TxOut>,
change_destination_script: Script,
feerate_sat_per_1000_weight: u32,
secp_ctx: &Secp256k1<C>
) -> Result<Transaction, ()>

See KeysManager::spend_spendable_outputs for documentation on this method.

pub fn derive_channel_keys(
&self,
channel_value_satoshis: u64,
params: &[u8; 32]
) -> InMemorySigner

See KeysManager::derive_channel_keys for documentation on this method.