Trait lightning::chain::Access

pub trait Access {
    fn get_utxo(
        &self,
        genesis_hash: &BlockHash,
        short_channel_id: u64
    ) -> Result<TxOut, AccessError>;
}

The Access trait defines behavior for accessing chain data and state, such as blocks and UTXOs.

Required Methods

fn get_utxo(
&self,
genesis_hash: &BlockHash,
short_channel_id: u64
) -> Result<TxOut, AccessError>

Returns the transaction output of a funding transaction encoded by short_channel_id. Returns an error if genesis_hash is for a different chain or if such a transaction output is unknown.