Trait lightning::chain::Listen

pub trait Listen {
    fn filtered_block_connected(
        &self,
        header: &BlockHeader,
        txdata: &TransactionData<'_>,
        height: u32
    );
    fn block_disconnected(&self, header: &BlockHeader, height: u32);

    fn block_connected(&self, block: &Block, height: u32) { ... }
}

The Listen trait is used to notify when blocks have been connected or disconnected from the chain.

Useful when needing to replay chain data upon startup or as new chain events occur. Clients sourcing chain data using a block-oriented API should prefer this interface over Confirm. Such clients fetch the entire header chain whereas clients using Confirm only fetch headers when needed.

By using Listen::filtered_block_connected this interface supports clients fetching the entire header chain and only blocks with matching transaction data using BIP 157 filters or other similar filtering.

Required Methods

fn filtered_block_connected(
&self,
header: &BlockHeader,
txdata: &TransactionData<'_>,
height: u32
)

Notifies the listener that a block was added at the given height, with the transaction data possibly filtered.

fn block_disconnected(&self, header: &BlockHeader, height: u32)

Notifies the listener that a block was removed at the given height.

Provided Methods

fn block_connected(&self, block: &Block, height: u32)

Notifies the listener that a block was added at the given height.