Struct lightning::ln::chan_utils::HolderCommitmentTransaction

pub struct HolderCommitmentTransaction {
    pub counterparty_sig: Signature,
    pub counterparty_htlc_sigs: Vec<Signature>,
    /* private fields */
}

Information needed to build and sign a holder’s commitment transaction.

The transaction is only signed once we are ready to broadcast.

Fields
counterparty_sig: Signature

Our counterparty’s signature for the transaction

counterparty_htlc_sigs: Vec<Signature>

All non-dust counterparty HTLC signatures, in the order they appear in the transaction

Implementations

impl HolderCommitmentTransaction

pub fn new(
commitment_tx: CommitmentTransaction,
counterparty_sig: Signature,
counterparty_htlc_sigs: Vec<Signature>,
holder_funding_key: &PublicKey,
counterparty_funding_key: &PublicKey
) -> Self

Create a new holder transaction with the given counterparty signatures. The funding keys are used to figure out which signature should go first when building the transaction for broadcast.