Struct lightning::ln::chan_utils::TrustedCommitmentTransaction

pub struct TrustedCommitmentTransaction<'a> { /* private fields */ }

A wrapper on CommitmentTransaction indicating that the derived fields (the built bitcoin transaction and the transaction creation keys) are trusted.

See trust() and verify() functions on CommitmentTransaction.

This structure implements Deref.

Implementations

impl<'a> TrustedCommitmentTransaction<'a>

pub fn txid(&self) -> Txid

The transaction ID of the built Bitcoin transaction

pub fn built_transaction(&self) -> &BuiltCommitmentTransaction

The pre-built Bitcoin commitment transaction

pub fn keys(&self) -> &TxCreationKeys

The pre-calculated transaction creation public keys.

pub fn opt_anchors(&self) -> bool

Should anchors be used.

pub fn get_htlc_sigs<T: Signing>(
&self,
htlc_base_key: &SecretKey,
channel_parameters: &DirectedChannelTransactionParameters<'_>,
secp_ctx: &Secp256k1<T>
) -> Result<Vec<Signature>, ()>

Get a signature for each HTLC which was included in the commitment transaction (ie for which HTLCOutputInCommitment::transaction_output_index.is_some()).

The returned Vec has one entry for each HTLC, and in the same order.

This function is only valid in the holder commitment context, it always uses EcdsaSighashType::All.