Function lightning::ln::inbound_payment::create_from_hash

pub fn create_from_hash(
    keys: &ExpandedKey,
    min_value_msat: Option<u64>,
    payment_hash: PaymentHash,
    invoice_expiry_delta_secs: u32,
    current_time: u64,
    min_final_cltv_expiry_delta: Option<u16>
) -> Result<PaymentSecret, ()>

Equivalent to crate::ln::channelmanager::ChannelManager::create_inbound_payment_for_hash, but no ChannelManager is required. Useful for generating invoices for phantom node payments without a ChannelManager.

See create for information on the keys and current_time parameters.

Note that if min_final_cltv_expiry_delta is set to some value, then the payment will not be receivable on versions of LDK prior to 0.0.114.