Enum lightning::ln::msgs::NetAddress

pub enum NetAddress {
    IPv4 {
        addr: [u8; 4],
        port: u16,
    },
    IPv6 {
        addr: [u8; 16],
        port: u16,
    },
    OnionV2([u8; 12]),
    OnionV3 {
        ed25519_pubkey: [u8; 32],
        checksum: u16,
        version: u8,
        port: u16,
    },
    Hostname {
        hostname: Hostname,
        port: u16,
    },
}

An address which can be used to connect to a remote peer.

Variants

IPv4

Fields

addr: [u8; 4]

The 4-byte IPv4 address

port: u16

The port on which the node is listening

An IPv4 address/port on which the peer is listening.

IPv6

Fields

addr: [u8; 16]

The 16-byte IPv6 address

port: u16

The port on which the node is listening

An IPv6 address/port on which the peer is listening.

OnionV2([u8; 12])

An old-style Tor onion address/port on which the peer is listening.

This field is deprecated and the Tor network generally no longer supports V2 Onion addresses. Thus, the details are not parsed here.

OnionV3

Fields

ed25519_pubkey: [u8; 32]

The ed25519 long-term public key of the peer

checksum: u16

The checksum of the pubkey and version, as included in the onion address

version: u8

The version byte, as defined by the Tor Onion v3 spec.

port: u16

The port on which the node is listening

A new-style Tor onion address/port on which the peer is listening.

To create the human-readable “hostname”, concatenate the ED25519 pubkey, checksum, and version, wrap as base32 and append “.onion”.

Hostname

Fields

hostname: Hostname

The hostname on which the node is listening.

port: u16

The port on which the node is listening.

A hostname/port on which the peer is listening.