Struct lightning::ln::peer_handler::IgnoringMessageHandler

pub struct IgnoringMessageHandler {}

A dummy struct which implements RoutingMessageHandler without storing any routing information or doing any processing. You can provide one of these as the route_handler in a MessageHandler.