Struct lightning::ln::script::ShutdownScript

pub struct ShutdownScript(_);

A script pubkey for shutting down a channel as defined by BOLT #2.

Implementations

impl ShutdownScript

pub fn new_p2wpkh(pubkey_hash: &WPubkeyHash) -> Self

Generates a P2WPKH script pubkey from the given WPubkeyHash.

pub fn new_p2wsh(script_hash: &WScriptHash) -> Self

Generates a P2WSH script pubkey from the given WScriptHash.

pub fn new_witness_program(
version: WitnessVersion,
program: &[u8]
) -> Result<Self, InvalidShutdownScript>

Generates a witness script pubkey from the given segwit version and program.

Note for version-zero witness scripts you must use ShutdownScript::new_p2wpkh or ShutdownScript::new_p2wsh instead.

Errors

This function may return an error if program is invalid for the segwit version.

pub fn into_inner(self) -> Script

Converts the shutdown script into the underlying Script.

pub fn as_legacy_pubkey(&self) -> Option<&PublicKey>

Returns the PublicKey used for a P2WPKH shutdown script if constructed directly from it.

pub fn is_compatible(&self, features: &InitFeatures) -> bool

Returns whether the shutdown script is compatible with the features as defined by BOLT #2.

Specifically, checks for compliance with feature option_shutdown_anysegwit.