Struct lightning::onion_message::OnionMessenger

pub struct OnionMessenger<ES: Deref, NS: Deref, L: Deref, CMH: Deref>where
    ES::Target: EntropySource,
    NS::Target: NodeSigner,
    L::Target: Logger,
    CMH::Target: CustomOnionMessageHandler,{ /* private fields */ }

A sender, receiver and forwarder of onion messages. In upcoming releases, this object will be used to retrieve invoices and fulfill invoice requests from offers. Currently, only sending and receiving custom onion messages is supported.

Example

// Create the onion messenger. This must use the same `keys_manager` as is passed to your
// ChannelManager.
let onion_messenger = OnionMessenger::new(&keys_manager, &keys_manager, logger, your_custom_message_handler);

impl Writeable for YourCustomMessage {
	fn write<W: Writer>(&self, w: &mut W) -> Result<(), io::Error> {
		// Write your custom onion message to `w`
	}
}
impl CustomOnionMessageContents for YourCustomMessage {
	fn tlv_type(&self) -> u64 {
		your_custom_message_type
	}
}
// Send a custom onion message to a node id.
let intermediate_hops = [hop_node_id1, hop_node_id2];
let reply_path = None;
let message = OnionMessageContents::Custom(your_custom_message);
onion_messenger.send_onion_message(&intermediate_hops, Destination::Node(destination_node_id), message, reply_path);

// Create a blinded path to yourself, for someone to send an onion message to.
let hops = [hop_node_id3, hop_node_id4, your_node_id];
let blinded_path = BlindedPath::new(&hops, &keys_manager, &secp_ctx).unwrap();

// Send a custom onion message to a blinded path.
let reply_path = None;
let message = OnionMessageContents::Custom(your_custom_message);
onion_messenger.send_onion_message(&intermediate_hops, Destination::BlindedPath(blinded_path), message, reply_path);

Implementations

impl<ES: Deref, NS: Deref, L: Deref, CMH: Deref> OnionMessenger<ES, NS, L, CMH>where
ES::Target: EntropySource,
NS::Target: NodeSigner,
L::Target: Logger,
CMH::Target: CustomOnionMessageHandler,

pub fn new(
entropy_source: ES,
node_signer: NS,
logger: L,
custom_handler: CMH
) -> Self

Constructs a new OnionMessenger to send, forward, and delegate received onion messages to their respective handlers.

pub fn send_onion_message<T: CustomOnionMessageContents>(
&self,
intermediate_nodes: &[PublicKey],
destination: Destination,
message: OnionMessageContents<T>,
reply_path: Option<BlindedPath>
) -> Result<(), SendError>

Send an onion message with contents message to destination, routing it through intermediate_nodes. See OnionMessenger for example usage.