Module lightning::routing::gossip

The top-level network map tracking logic lives here.

Structs

ChannelInfo

Details about a channel (both directions). Received within a channel announcement.

ChannelUpdateInfo

Details about one direction of a channel as received within a ChannelUpdate.

DirectedChannelInfo

A wrapper around ChannelInfo representing information about the channel as directed from a source node to a target node.

NetworkGraph

Represents the network as nodes and channels between them

NodeAlias

A user-defined name for a node, which may be used when displaying the node in a graph.

NodeAnnouncementInfo

Information received in the latest node_announcement from this node.

NodeId

Represents the compressed public key of a node

NodeInfo

Details about a node in the network, known from the network announcement.

P2PGossipSync

Receives and validates network updates from peers, stores authentic and relevant data as a network graph. This network graph is then used for routing payments. Provides interface to help with initial routing sync by serving historical announcements.

ReadOnlyNetworkGraph

A read-only view of NetworkGraph.

RoutingFees

Fees for routing via a given channel or a node

Enums

EffectiveCapacity

The effective capacity of a channel for routing purposes.

NetworkUpdate

Update to the NetworkGraph based on payment failure information conveyed via the Onion return packet by a node along the route. See BOLT #4 for details.

Constants

UNKNOWN_CHANNEL_CAPACITY_MSAT

The presumed channel capacity denominated in millisatoshi for EffectiveCapacity::Unknown to use when making routing decisions.