Struct lightning::routing::gossip::NetworkGraph

pub struct NetworkGraph<L: Deref>where
    L::Target: Logger,{ /* private fields */ }

Represents the network as nodes and channels between them

Implementations

impl<L: Deref> NetworkGraph<L>where
L::Target: Logger,

pub fn handle_network_update(&self, network_update: &NetworkUpdate)

Handles any network updates originating from Events.

impl<L: Deref> NetworkGraph<L>where
L::Target: Logger,

pub fn new(genesis_hash: BlockHash, logger: L) -> NetworkGraph<L>

Creates a new, empty, network graph.

pub fn read_only(&self) -> ReadOnlyNetworkGraph<'_>

Returns a read-only view of the network graph.

pub fn get_last_rapid_gossip_sync_timestamp(&self) -> Option<u32>

The unix timestamp provided by the most recent rapid gossip sync. It will be set by the rapid sync process after every sync completion.

pub fn set_last_rapid_gossip_sync_timestamp(
&self,
last_rapid_gossip_sync_timestamp: u32
)

Update the unix timestamp provided by the most recent rapid gossip sync. This should be done automatically by the rapid sync process after every sync completion.

pub fn update_node_from_announcement(
&self,
msg: &NodeAnnouncement
) -> Result<(), LightningError>

For an already known node (from channel announcements), update its stored properties from a given node announcement.

You probably don’t want to call this directly, instead relying on a P2PGossipSync’s RoutingMessageHandler implementation to call it indirectly. This may be useful to accept routing messages from a source using a protocol other than the lightning P2P protocol.

pub fn update_node_from_unsigned_announcement(
&self,
msg: &UnsignedNodeAnnouncement
) -> Result<(), LightningError>

For an already known node (from channel announcements), update its stored properties from a given node announcement without verifying the associated signatures. Because we aren’t given the associated signatures here we cannot relay the node announcement to any of our peers.

pub fn update_channel_from_announcement<C: Deref>(
&self,
msg: &ChannelAnnouncement,
chain_access: &Option<C>
) -> Result<(), LightningError>where
C::Target: Access,

Store or update channel info from a channel announcement.

You probably don’t want to call this directly, instead relying on a P2PGossipSync’s RoutingMessageHandler implementation to call it indirectly. This may be useful to accept routing messages from a source using a protocol other than the lightning P2P protocol.

If a chain::Access object is provided via chain_access, it will be called to verify the corresponding UTXO exists on chain and is correctly-formatted.

pub fn update_channel_from_unsigned_announcement<C: Deref>(
&self,
msg: &UnsignedChannelAnnouncement,
chain_access: &Option<C>
) -> Result<(), LightningError>where
C::Target: Access,

Store or update channel info from a channel announcement without verifying the associated signatures. Because we aren’t given the associated signatures here we cannot relay the channel announcement to any of our peers.

If a chain::Access object is provided via chain_access, it will be called to verify the corresponding UTXO exists on chain and is correctly-formatted.

pub fn add_channel_from_partial_announcement(
&self,
short_channel_id: u64,
timestamp: u64,
features: ChannelFeatures,
node_id_1: PublicKey,
node_id_2: PublicKey
) -> Result<(), LightningError>

Update channel from partial announcement data received via rapid gossip sync

timestamp: u64: Timestamp emulating the backdated original announcement receipt (by the rapid gossip sync server)

All other parameters as used in msgs::UnsignedChannelAnnouncement fields.

pub fn channel_failed(&self, short_channel_id: u64, is_permanent: bool)

Marks a channel in the graph as failed if a corresponding HTLC fail was sent. If permanent, removes a channel from the local storage. May cause the removal of nodes too, if this was their last channel. If not permanent, makes channels unavailable for routing.

pub fn node_failed_permanent(&self, node_id: &PublicKey)

Marks a node in the graph as permanently failed, effectively removing it and its channels from local storage.

pub fn remove_stale_channels_and_tracking(&self)

Removes information about channels that we haven’t heard any updates about in some time. This can be used regularly to prune the network graph of channels that likely no longer exist.

While there is no formal requirement that nodes regularly re-broadcast their channel updates every two weeks, the non-normative section of BOLT 7 currently suggests that pruning occur for updates which are at least two weeks old, which we implement here.

Note that for users of the lightning-background-processor crate this method may be automatically called regularly for you.

This method will also cause us to stop tracking removed nodes and channels if they have been in the map for a while so that these can be resynced from gossip in the future.

This method is only available with the std feature. See NetworkGraph::remove_stale_channels_and_tracking_with_time for no-std use.

pub fn remove_stale_channels_and_tracking_with_time(
&self,
current_time_unix: u64
)

Removes information about channels that we haven’t heard any updates about in some time. This can be used regularly to prune the network graph of channels that likely no longer exist.

While there is no formal requirement that nodes regularly re-broadcast their channel updates every two weeks, the non-normative section of BOLT 7 currently suggests that pruning occur for updates which are at least two weeks old, which we implement here.

This method will also cause us to stop tracking removed nodes and channels if they have been in the map for a while so that these can be resynced from gossip in the future.

This function takes the current unix time as an argument. For users with the std feature enabled, NetworkGraph::remove_stale_channels_and_tracking may be preferable.

pub fn update_channel(&self, msg: &ChannelUpdate) -> Result<(), LightningError>

For an already known (from announcement) channel, update info about one of the directions of the channel.

You probably don’t want to call this directly, instead relying on a P2PGossipSync’s RoutingMessageHandler implementation to call it indirectly. This may be useful to accept routing messages from a source using a protocol other than the lightning P2P protocol.

If built with no-std, any updates with a timestamp more than two weeks in the past or materially in the future will be rejected.

pub fn update_channel_unsigned(
&self,
msg: &UnsignedChannelUpdate
) -> Result<(), LightningError>

For an already known (from announcement) channel, update info about one of the directions of the channel without verifying the associated signatures. Because we aren’t given the associated signatures here we cannot relay the channel update to any of our peers.

If built with no-std, any updates with a timestamp more than two weeks in the past or materially in the future will be rejected.