Struct lightning::routing::gossip::P2PGossipSync

pub struct P2PGossipSync<G: Deref<Target = NetworkGraph<L>>, C: Deref, L: Deref>where
    C::Target: Access,
    L::Target: Logger,{ /* private fields */ }

Receives and validates network updates from peers, stores authentic and relevant data as a network graph. This network graph is then used for routing payments. Provides interface to help with initial routing sync by serving historical announcements.

Implementations

impl<G: Deref<Target = NetworkGraph<L>>, C: Deref, L: Deref> P2PGossipSync<G, C, L>where
C::Target: Access,
L::Target: Logger,

pub fn new(network_graph: G, chain_access: Option<C>, logger: L) -> Self

Creates a new tracker of the actual state of the network of channels and nodes, assuming an existing Network Graph. Chain monitor is used to make sure announced channels exist on-chain, channel data is correct, and that the announcement is signed with channel owners’ keys.

pub fn add_chain_access(&mut self, chain_access: Option<C>)

Adds a provider used to check new announcements. Does not affect existing announcements unless they are updated. Add, update or remove the provider would replace the current one.

pub fn network_graph(&self) -> &G

Gets a reference to the underlying NetworkGraph which was provided in P2PGossipSync::new.

(C-not exported) as bindings don’t support a reference-to-a-reference yet