Module lightning::routing::router

The top-level routing/network map tracking logic lives here.

You probably want to create a P2PGossipSync and use that as your RoutingMessageHandler and then interrogate it to get routes for your own payments.

Structs

DefaultRouter

A Router implemented using find_route.

InFlightHtlcs

A data structure for tracking in-flight HTLCs. May be used during pathfinding to account for in-use channel liquidity.

PaymentParameters

The recipient of a payment.

Route

A route directs a payment from the sender (us) to the recipient. If the recipient supports MPP, it can take multiple paths. Each path is composed of one or more hops through the network.

RouteHint

A list of hops along a payment path terminating with a channel to the recipient.

RouteHintHop

A channel descriptor for a hop along a payment path.

RouteHop

A hop in a route

RouteParameters

Parameters needed to find a Route.

ScorerAccountingForInFlightHtlcs

Score implementation that factors in in-flight HTLC liquidity.

Constants

DEFAULT_MAX_PATH_COUNT

Maximum number of paths we allow an (MPP) payment to have.

DEFAULT_MAX_TOTAL_CLTV_EXPIRY_DELTA

Maximum total CTLV difference we allow for a full payment path.

Traits

Router

A trait defining behavior for routing a payment.

Functions

build_route_from_hops

Construct a route from us (payer) to the target node (payee) via the given hops (which should exclude the payer, but include the payee). This may be useful, e.g., for probing the chosen path.

find_route

Finds a route from us (payer) to the given target node (payee).