Function lightning::routing::router::build_route_from_hops

pub fn build_route_from_hops<L: Deref, GL: Deref>(
    our_node_pubkey: &PublicKey,
    hops: &[PublicKey],
    route_params: &RouteParameters,
    network_graph: &NetworkGraph<GL>,
    logger: L,
    random_seed_bytes: &[u8; 32]
) -> Result<Route, LightningError>where
    L::Target: Logger,
    GL::Target: Logger,

Construct a route from us (payer) to the target node (payee) via the given hops (which should exclude the payer, but include the payee). This may be useful, e.g., for probing the chosen path.

Re-uses logic from find_route, so the restrictions described there also apply here.