Trait lightning::routing::router::Router

pub trait Router {
    fn find_route(
        &self,
        payer: &PublicKey,
        route_params: &RouteParameters,
        first_hops: Option<&[&ChannelDetails]>,
        inflight_htlcs: &InFlightHtlcs
    ) -> Result<Route, LightningError>;
    fn notify_payment_path_failed(
        &self,
        path: &[&RouteHop],
        short_channel_id: u64
    );
    fn notify_payment_path_successful(&self, path: &[&RouteHop]);
    fn notify_payment_probe_successful(&self, path: &[&RouteHop]);
    fn notify_payment_probe_failed(
        &self,
        path: &[&RouteHop],
        short_channel_id: u64
    );

    fn find_route_with_id(
        &self,
        payer: &PublicKey,
        route_params: &RouteParameters,
        first_hops: Option<&[&ChannelDetails]>,
        inflight_htlcs: &InFlightHtlcs,
        _payment_hash: PaymentHash,
        _payment_id: PaymentId
    ) -> Result<Route, LightningError> { ... }
}

A trait defining behavior for routing a payment.

Required Methods

fn find_route(
&self,
payer: &PublicKey,
route_params: &RouteParameters,
first_hops: Option<&[&ChannelDetails]>,
inflight_htlcs: &InFlightHtlcs
) -> Result<Route, LightningError>

Finds a Route between payer and payee for a payment with the given values.

fn notify_payment_path_failed(&self, path: &[&RouteHop], short_channel_id: u64)

Lets the router know that payment through a specific path has failed.

fn notify_payment_path_successful(&self, path: &[&RouteHop])

Lets the router know that payment through a specific path was successful.

fn notify_payment_probe_successful(&self, path: &[&RouteHop])

Lets the router know that a payment probe was successful.

fn notify_payment_probe_failed(&self, path: &[&RouteHop], short_channel_id: u64)

Lets the router know that a payment probe failed.

Provided Methods

fn find_route_with_id(
&self,
payer: &PublicKey,
route_params: &RouteParameters,
first_hops: Option<&[&ChannelDetails]>,
inflight_htlcs: &InFlightHtlcs,
_payment_hash: PaymentHash,
_payment_id: PaymentId
) -> Result<Route, LightningError>

Finds a Route between payer and payee for a payment with the given values. Includes PaymentHash and PaymentId to be able to correlate the request with a specific payment.