Module lightning::routing::scoring

Utilities for scoring payment channels.

ProbabilisticScorer may be given to find_route to score payment channels during path finding when a custom Score implementation is not needed.

Example

// Use the default channel penalties.
let params = ProbabilisticScoringParameters::default();
let scorer = ProbabilisticScorer::new(params, &network_graph, &logger);

// Or use custom channel penalties.
let params = ProbabilisticScoringParameters {
    liquidity_penalty_multiplier_msat: 2 * 1000,
    ..ProbabilisticScoringParameters::default()
};
let scorer = ProbabilisticScorer::new(params, &network_graph, &logger);

let route = find_route(&payer, &route_params, &network_graph, None, &logger, &scorer, &random_seed_bytes);

Note

Persisting when built with feature no-std and restoring without it, or vice versa, uses different types and thus is undefined.

Structs

ChannelUsage

Proposed use of a channel passed as a parameter to Score::channel_penalty_msat.

FixedPenaltyScorer

Score implementation that uses a fixed penalty.

ProbabilisticScorerUsingTime

Probabilistic Score implementation.

ProbabilisticScoringParameters

Parameters for configuring ProbabilisticScorer.

Traits

LockableScore

A scorer that is accessed under a lock.

Score

An interface used to score payment channels for path finding.

WriteableScore

Refers to a scorer that is accessible under lock and also writeable to disk

Type Definitions

ProbabilisticScorer

Score implementation using channel success probability distributions.