Trait lightning::routing::scoring::LockableScore

pub trait LockableScore<'a> {
    type Locked: 'a + Score;

    fn lock(&'a self) -> Self::Locked;
}

A scorer that is accessed under a lock.

Needed so that calls to Score::channel_penalty_msat in find_route can be made while having shared ownership of a scorer but without requiring internal locking in Score implementations. Internal locking would be detrimental to route finding performance and could result in Score::channel_penalty_msat returning a different value for the same channel.

Required Associated Types

type Locked: 'a + Score

The locked Score type.

Required Methods

fn lock(&'a self) -> Self::Locked

Returns the locked scorer.