Trait lightning::util::events::MessageSendEventsProvider

pub trait MessageSendEventsProvider {
    fn get_and_clear_pending_msg_events(&self) -> Vec<MessageSendEvent>;
}

A trait indicating an object may generate message send events

Required Methods

fn get_and_clear_pending_msg_events(&self) -> Vec<MessageSendEvent>

Gets the list of pending events which were generated by previous actions, clearing the list in the process.