Trait lightning::util::events::OnionMessageProvider

pub trait OnionMessageProvider {
    fn next_onion_message_for_peer(
        &self,
        peer_node_id: PublicKey
    ) -> Option<OnionMessage>;
}

A trait indicating an object may generate onion messages to send

Required Methods

fn next_onion_message_for_peer(
&self,
peer_node_id: PublicKey
) -> Option<OnionMessage>

Gets the next pending onion message for the peer with the given node id.