Module lightning::util::persist

This module contains a simple key-value store trait KVStorePersister that allows one to implement the persistence for ChannelManager, NetworkGraph, and ChannelMonitor all in one place.

Traits

KVStorePersister

Trait for a key-value store for persisting some writeable object at some key Implementing KVStorePersister provides auto-implementations for Persister and Persist traits. It uses “manager”, “network_graph”, and “monitors/{funding_txo_id}_{funding_txo_index}” for keys.

Persister

Trait that handles persisting a ChannelManager, NetworkGraph, and WriteableScore to disk.