Trait lightning::util::persist::Persister

pub trait Persister<'a, M: Deref, T: Deref, ES: Deref, NS: Deref, SP: Deref, F: Deref, R: Deref, L: Deref, S: WriteableScore<'a>>where
    M::Target: 'static + Watch<<SP::Target as SignerProvider>::Signer>,
    T::Target: 'static + BroadcasterInterface,
    ES::Target: 'static + EntropySource,
    NS::Target: 'static + NodeSigner,
    SP::Target: 'static + SignerProvider,
    F::Target: 'static + FeeEstimator,
    R::Target: 'static + Router,
    L::Target: 'static + Logger,{
    fn persist_manager(
        &self,
        channel_manager: &ChannelManager<M, T, ES, NS, SP, F, R, L>
    ) -> Result<(), Error>;
    fn persist_graph(&self, network_graph: &NetworkGraph<L>) -> Result<(), Error>;
    fn persist_scorer(&self, scorer: &S) -> Result<(), Error>;
}

Trait that handles persisting a ChannelManager, NetworkGraph, and WriteableScore to disk.

Required Methods

fn persist_manager(
&self,
channel_manager: &ChannelManager<M, T, ES, NS, SP, F, R, L>
) -> Result<(), Error>

Persist the given [‘ChannelManager’] to disk, returning an error if persistence failed.

fn persist_graph(&self, network_graph: &NetworkGraph<L>) -> Result<(), Error>

Persist the given NetworkGraph to disk, returning an error if persistence failed.

fn persist_scorer(&self, scorer: &S) -> Result<(), Error>

Persist the given WriteableScore to disk, returning an error if persistence failed.