Trait lightning::util::ser::ReadableArgs

pub trait ReadableArgs<P>where
    Self: Sized,{
    fn read<R: Read>(reader: &mut R, params: P) -> Result<Self, DecodeError>;
}

A trait that various higher-level LDK types implement allowing them to be read in from a Read given some additional set of arguments which is required to deserialize.

(C-not exported) as we only export serialization to/from byte arrays instead

Required Methods

fn read<R: Read>(reader: &mut R, params: P) -> Result<Self, DecodeError>

Reads a Self in from the given Read.