Trait lightning::util::ser::Writeable

pub trait Writeable {
    fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error>;

    fn encode(&self) -> Vec<u8> { ... }
    fn serialized_length(&self) -> usize { ... }
}

A trait that various LDK types implement allowing them to be written out to a Writer.

(C-not exported) as we only export serialization to/from byte arrays instead

Required Methods

fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error>

Writes self out to the given Writer.

Provided Methods

fn encode(&self) -> Vec<u8>

Writes self out to a Vec<u8>.

fn serialized_length(&self) -> usize

Gets the length of this object after it has been serialized. This can be overridden to optimize cases where we prepend an object with its length.